client/requirements.txt
requests>=2.31.0
psutil>=5.9.8
py-cpuinfo>=9.0.0
GPUtil>=1.4.0
tqdm>=4.66.4
